+++ test/unit/fuzz-tools.test.js
@@ -483,6 +483,7 @@ import * as metavirFibrosisV504 from '../../lib/metavir-fibrosis-v504.js';
 import * as metavirActivityV505 from '../../lib/metavir-activity-v505.js';
 import * as jergerTympanogramV506 from '../../lib/jerger-tympanogram-v506.js';
 import * as vhi10V508 from '../../lib/vhi10-v508.js';
+import * as sunnybrookFacialV509 from '../../lib/sunnybrook-facial-v509.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -911,6 +912,7 @@ const MODULES = {
   'metavir-activity-v505.js': metavirActivityV505,
   'jerger-tympanogram-v506.js': jergerTympanogramV506,
   'vhi10-v508.js': vhi10V508,
+  'sunnybrook-facial-v509.js': sunnybrookFacialV509,
   'tb-testing.js': tbTesting,
 };